Refatoração na validação do carrinho

+++ NerdStoreEnterprise/src/services/NSE.Carrinho.API/Controllers/CarrinhoController.cs
@@ -6,6 +6,7 @@ using NSE.Carrinho.API.Model;
 using NSE.WebAPI.Core.Controllers;
 using NSE.WebAPI.Core.Usuario;
 using System;
+using System.Linq;
 using System.Threading.Tasks;
 
 namespace NSE.Carrinho.API.Controllers
@@ -38,6 +39,7 @@ namespace NSE.Carrinho.API.Controllers
             else
                 ManipularCarrinhoExistente(carrinho, item);
 
+            ValidarCarrinho(carrinho);
             if (!OperacaoValida())
                 return CustomResponse();
 
@@ -83,6 +85,10 @@ namespace NSE.Carrinho.API.Controllers
 
             carrinho.AtualizarUnidades(itemCarrinho, item.Quantidade);
 
+            ValidarCarrinho(carrinho);
+            if (!OperacaoValida())
+                return CustomResponse();
+
             _context.CarrinhoItens.Update(itemCarrinho);
             _context.CarrinhoCliente.Update(carrinho);
 
@@ -98,6 +104,12 @@ namespace NSE.Carrinho.API.Controllers
             var itemCarrinho = await ObterItemCarrinhoValidado(produtoId, carrinho);
             if (itemCarrinho == null)
                 return CustomResponse();
+
+            ValidarCarrinho(carrinho);
+            if (!OperacaoValida())
+                return CustomResponse();
+
+            carrinho.RemoverItem(itemCarrinho);
 
             _context.CarrinhoItens.Remove(itemCarrinho);
             _context.CarrinhoCliente.Update(carrinho);
@@ -144,6 +156,15 @@ namespace NSE.Carrinho.API.Controllers
             var result = await _context.SaveChangesAsync();
             if (result <= 0)
                 AdicionarErroProcessamento("Não foi possível persistir os dados no banco");
+        }
+
+        private bool ValidarCarrinho(CarrinhoCliente carrinho)
+        {
+            if (carrinho.EhValido())
+                return true;
+
+            carrinho.ValidationResult.Errors.ToList().ForEach(e => AdicionarErroProcessamento(e.ErrorMessage));
+            return false;
         }
     }
 }
